fix non-normal fixtures moved forward in time (#102)
fixture between aston villa and liverpool was assigned to GW 29, but was played recently -- way before gw 29 was set to start. I didn't anticipate this ever happening, only that fixtures would be moved back, and so the GW28 fixtures were wiped out of being normal.

This resulted in users being unable to place a bet for any fixtures for GW 28 which just opened up, as no GW 28 fixture was set as normal.